http-service/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
itsdangerous==1.1.0
jaraco.classes==3.2.0
jaraco.collections==3.1.0
jaraco.functools==3.1.0
jaraco.text==3.4.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.6.0
mysql-connector-python==8.0.23
portend==2.7.0
protobuf==3.14.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
tempora==4.0.1
urllib3==1.26.3
Werkzeug==1.0.1
zc.lockfile==2.0
